more parametrization of inputlayer
