+++ src/destin/__init__.py
@@ -4,7 +4,6 @@ from .input import OpenCVInputLayer
 from .input import ROSInputLayer
 from .nodes import AutoEncoderNode
 from .nodes import StackedAutoEncoderNode
-from .architectures import NetworkArchitecture
-from .architectures import DestinArchitecture
+from .architectures import *
 
 #from ..destin.msg import DestinNodeState
