+++ src/destin/architectures/__init__.py
@@ -1,2 +1,3 @@
 from .architecture import NetworkArchitecture
 from .destin_arch import DestinArchitecture
+from .handcoded_destin import HandcodedDestinArchitecture
